+++ src/app/index.ts
@@ -1,0 +1,3 @@
+import { buildApp } from "./app";
+
+export { buildApp };
